+++ src/main/java/org/example/outsourcing_project/domain/auth/dto/SignupRequestDto.java
@@ -1,4 +1,4 @@
-package org.example.outsourcing_project.domain.user.dto.request;
+package org.example.outsourcing_project.domain.auth.dto;
 
 import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
@@ -7,7 +7,7 @@ import org.example.outsourcing_project.domain.user.UserRole;
 
 @Getter
 @AllArgsConstructor
-public class SignupUserRequestDto {
+public class SignupRequestDto {
 
     // @Email 은 간단하지만 빈 문자열을 통과 시킨다.(@Pattern을 통한 정규식 검사를 더 많이 사용)
     @Pattern(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,6}$", message = "이메일 형식이 올바르지 않습니다.")
