+++ src/main/java/org/example/outsourcing_project/domain/user/controller/UserController.java
@@ -1,10 +1,8 @@
 package org.example.outsourcing_project.domain.user.controller;
 
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.example.outsourcing_project.common.jwt.LoginUser;
 import org.example.outsourcing_project.domain.user.dto.request.DeleteUserRequestDto;
-import org.example.outsourcing_project.domain.user.dto.request.SignupUserRequestDto;
 import org.example.outsourcing_project.domain.user.dto.request.UpdateUserRequestDto;
 import org.example.outsourcing_project.domain.user.dto.response.UserResponseDto;
 import org.example.outsourcing_project.domain.user.service.UserService;
@@ -20,15 +18,6 @@ import java.util.List;
 public class UserController {
 
     private final UserService userService;
-
-    // 회원가입
-    @PostMapping("/signup")
-    public ResponseEntity<UserResponseDto> Signup(@Valid @RequestBody SignupUserRequestDto requestDto) {
-
-        UserResponseDto responseDto = userService.Signup(requestDto.getEmail(), requestDto.getPassword(), requestDto.getName(), requestDto.getAddress(), requestDto.getRole());
-
-        return new ResponseEntity<>(responseDto, HttpStatus.CREATED);
-    }
 
     // 전체 유저 조회
     @GetMapping
